added add_task

+++ core/models.py
@@ -3,4 +3,4 @@ from django.db import models
 
 class Task(models.Model):
     task = models.TextField()
-    owner_id = models.IntegerField()
+    owner_id = models.BigIntegerField()
